chore: upgrade angular core, material, cli and cdk from version 13.2.0 to version 19.1.3 and typescript from 4.5.5 to 5.7.3

+++ src/app/modules/dashboard/components/parcel-item/parcel-item.component.ts
@@ -2,15 +2,16 @@ import {Component, EventEmitter, Input, OnInit, Output} from '@angular/core';
 import {Parcel} from '../../../../shared/models/parcel';
 import {ParcelStatusEnum} from '../../../../shared/models/parcel-status-enum';
 import {
-  faShoppingBag,
-  faTruck,
-  faClock,
-  faShareAlt,
-  faTrash,
   faBoxOpen,
-  faShippingFast,
+  faClipboardList,
+  faClock,
   faMoneyCheckAlt,
-  faPen, faClipboardList
+  faPen,
+  faShareAlt,
+  faShippingFast,
+  faShoppingBag,
+  faTrash,
+  faTruck
 } from '@fortawesome/free-solid-svg-icons';
 import {DeleteDialogComponent} from '../../../../shared/components/dialogs/delete-dialog/delete-dialog.component';
 import {ParcelService} from '../../../../shared/services/parcel/parcel.service';
@@ -18,19 +19,32 @@ import {DashboardLoadingService} from '../../pages/dashboard/dashboard-loading.s
 import {withLoading} from '../../../../shared/helpers/operators';
 import {Router} from '@angular/router';
 import {EDIT_PARCEL} from '../../../../shared/constants/endpoints';
-import {prefixUrl} from '../../../../shared/helpers/url.helper';
+import {UrlHelperService} from '../../../../shared/services/url/url.helper.service';
 import {MatDialog} from '@angular/material/dialog';
 import {MatSnackBar} from '@angular/material/snack-bar';
 import {Clipboard} from '@angular/cdk/clipboard';
+import {DatePipe, NgClass, NgIf, NgSwitch, NgSwitchCase} from '@angular/common';
+import {FontAwesomeModule} from '@fortawesome/angular-fontawesome';
+import {MatTooltip} from '@angular/material/tooltip';
 
 @Component({
   selector: 'app-parcel-item',
   templateUrl: './parcel-item.component.html',
-  styleUrls: ['./parcel-item.component.scss']
+  styleUrls: ['./parcel-item.component.scss'],
+  imports: [
+    NgIf,
+    NgClass,
+    FontAwesomeModule,
+    NgSwitchCase,
+    NgSwitch,
+    MatTooltip,
+    DatePipe
+  ],
+  standalone: true
 })
 export class ParcelItemComponent implements OnInit {
 
-  @Input() parcel: Parcel;
+  @Input() parcel?: Parcel;
   @Output() parcelDeleted = new EventEmitter<Parcel>();
   parcelStatuses = ParcelStatusEnum;
 
@@ -57,57 +71,65 @@ export class ParcelItemComponent implements OnInit {
   };
 
   constructor(private dialog: MatDialog, private parcelService: ParcelService, private dashboardLoadingService: DashboardLoadingService,
-              private router: Router, private snackBar: MatSnackBar, private clipboard: Clipboard) {
+              private router: Router, private snackBar: MatSnackBar, private clipboard: Clipboard, private urlHelper: UrlHelperService) {
   }
 
   ngOnInit() {
   }
 
   onParcelDeleteClick() {
-    const dialogRef = this.dialog.open(DeleteDialogComponent, {
-      data: {toDelete: this.parcel.title},
-      panelClass: 'app-dialog'
-    });
+    if (this.parcel) {
+      const dialogRef = this.dialog.open(DeleteDialogComponent, {
+        data: {toDelete: this.parcel?.title},
+        panelClass: 'app-dialog'
+      });
 
-    dialogRef.afterClosed().subscribe(result => {
-      if (result) {
-        this.deleteParcel();
-      }
-    });
+      dialogRef.afterClosed().subscribe(result => {
+        if (result) {
+          this.deleteParcel();
+        }
+      });
+    }
   }
 
   deleteParcel() {
-    this.parcelService.deleteParcel(this.parcel.id)
-      .pipe(withLoading(this.dashboardLoadingService.loading$))
-      .subscribe(value => {
-        this.parcelDeleted.emit(this.parcel);
-      }, error => {
-      });
+    if (this.parcel && this.parcel.id) {
+      this.parcelService.deleteParcel(this.parcel.id)
+        .pipe(withLoading(this.dashboardLoadingService.loading$))
+        .subscribe(value => {
+          this.parcelDeleted.emit(this.parcel);
+        }, error => {
+        });
+    }
   }
 
   onEditParcelClick() {
-    this.router.navigateByUrl(EDIT_PARCEL(this.parcel.id));
+    if (this.parcel && this.parcel.id) this.router.navigateByUrl(EDIT_PARCEL(this.parcel.id));
   }
 
   onShareParcelClick() {
-    const trackingUrl = this.parcel.trackingUrl;
-    // Construct the snackbar message and copy the tracking url if its present.
-    let message;
-    if (trackingUrl) {
-      const copied = this.clipboard.copy(trackingUrl);
-      message = copied ? this.copiedSuccessMsg : this.copiedErrorMsg;
-    } else {
-      message = this.copiedMissingUrlMsg;
-    }
+    if (this.parcel) {
+      const trackingUrl = this.parcel.trackingUrl;
+      // Construct the snackbar message and copy the tracking url if its present.
+      let message;
+      if (trackingUrl) {
+        const copied = this.clipboard.copy(trackingUrl);
+        message = copied ? this.copiedSuccessMsg : this.copiedErrorMsg;
+      } else {
+        message = this.copiedMissingUrlMsg;
+      }
 
-    // Display the snackbar message.
-    this.snackBar.open(message, 'OK', {duration: 2500});
+      // Display the snackbar message.
+      this.snackBar.open(message, 'OK', {duration: 2500});
+    }
   }
 
   goToTrackingUrl() {
-    const trackingUrl = this.parcel.trackingUrl;
-    if (trackingUrl) {
-      window.open(prefixUrl(trackingUrl), '_blank');
+    if (this.parcel) {
+      const trackingUrl = this.parcel.trackingUrl;
+      if (trackingUrl) {
+        window.open(this.urlHelper.prefixUrl(trackingUrl), '_blank');
+      }
     }
   }
